Switch submitted_by with a user field
Use a foreign key field to store the user from kaznet.apps.users instead
of storing the username of the submitter. submitted_by was a CharField
storing the username but it's better to have foreign key to the user.

Users who submit should always have accounts in kaznet else we ignore the
submission assuming it was erroneous.

Signed-off-by: Njagi Mwaniki <njagi@urbanslug.com>